reviewpro_pipeline/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
rapidfuzz>=3.0.0
python-dateutil>=2.8.0
xlsxwriter>=3.1.0
